requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
python-multipart

